Move Readable implementations for marker types
They are now implemented at the same position they are declared.
